+++ examples/streamz/scripts/entry.sh
@@ -121,10 +121,10 @@ log "INFO" "Sample data read into kafka topic, $input_topic"
 #**********************************
 log "INFO" "Preparing to run cybert"
 if [ ! -z "$benchmark" ] ; then
-  log "INFO" "python -i $CYBERT_HOME/python/cybert.py --input_topic input --output_topic output --group_id $group_id --model $model_file --label_map $label_file --cuda_visible_devices $cuda_visible_devices --poll_interval $poll_interval --max_batch_size $max_batch_size --benchmark $benchmark"
-  python -i $CYBERT_HOME/python/cybert.py --input_topic input --output_topic output --group_id $group_id --model $model_file --label_map $label_file --poll_interval $poll_interval --max_batch_size $max_batch_size --benchmark $benchmark
+  log "INFO" "python -i $CYBERT_HOME/python/cybert.py --input_topic $input_topic --output_topic $output_topic --group_id $group_id --model $model_file --label_map $label_file --cuda_visible_devices $cuda_visible_devices --poll_interval $poll_interval --max_batch_size $max_batch_size --benchmark $benchmark"
+  python -i $CYBERT_HOME/python/cybert.py --input_topic $input_topic --output_topic $output_topic --group_id $group_id --model $model_file --label_map $label_file --poll_interval $poll_interval --max_batch_size $max_batch_size --benchmark $benchmark
 else
-  log "INFO" "python -i $CYBERT_HOME/python/cybert.py --input_topic input --output_topic output --group_id $group_id --model $model_file --label_map $label_file --cuda_visible_devices $cuda_visible_devices --poll_interval $poll_interval --max_batch_size $max_batch_size"
-  python -i $CYBERT_HOME/python/cybert.py --input_topic input --output_topic output --group_id $group_id --model $model_file --label_map $label_file --poll_interval $poll_interval --max_batch_size $max_batch_size
+  log "INFO" "python -i $CYBERT_HOME/python/cybert.py --input_topic $input_topic --output_topic $output_topic --group_id $group_id --model $model_file --label_map $label_file --cuda_visible_devices $cuda_visible_devices --poll_interval $poll_interval --max_batch_size $max_batch_size"
+  python -i $CYBERT_HOME/python/cybert.py --input_topic $input_topic --output_topic $output_topic --group_id $group_id --model $model_file --label_map $label_file --poll_interval $poll_interval --max_batch_size $max_batch_size
 fi